add anomaly_handler to fix issue with logging anomalies without anomaly_db
